refactor: 统一时间戳/duration/timeout 类型 (v3.9.8)
按用户要求"找找函数定义不一致, 一边 number 一边 string". 用 AST 扫描
全仓 366 .py 文件 -> 39 处 root-level 类型冲突. 3 个真有 bug 高发区:

1. timestamp 字段 (created_at / started_at / finished_at)
   - 18 个 dataclass 用 str (ISO-8601), 但 ApprovalRequest / ActionResult
     用 float (Unix epoch). 同字段跨 API 类型不同, 前端时间戳拼接会 throw.

2. duration_ms
   - ToolResult.duration_ms / executor._failed_result 用 int
   - RuntimeStep.duration_ms / TrajectoryRecord / TrajectoryMetrics /
     context_compactor 4 处用 float (sub-ms precision 无意义)

3. timeout
   - subscribe / wait / recv 用 float
   - _run_git / execute_python_code / _run_shell 用 int

修复方向: 一律向大头靠, 全部统一:
- timestamp: str (ISO-8601 UTC), agent.task.Task / dispatcher / store 全部改
- duration_ms: int (亚毫秒精度没用, 整型毫秒覆盖 ~2.9e6 小时无溢出)
- timeout: int (秒), wait/subscribe/recv 都改

新增 agent/runtime/utils/time_utils.py: to_iso/from_iso/now_iso/
duration_ms 统一转换工具.

修改文件清单:

+ agent/runtime/utils/{__init__,time_utils}.py  新建
M agent/approval.py
  - ApprovalRequest.created_at / resolved_at: float -> str (ISO)
  - ApprovalStore.wait.timeout: float -> int
  - ApprovalStore.get_pending 用 datetime 比较替代 time.time() - created_at
  - ApprovalStore._load_history 接 JSONL 兼容 float/epoch 和 str/ISO
  - _to_dict created_at / created_at_iso 同步

M agent/runtime/actions/models.py
  - ActionResult.started_at / finished_at: float -> str
  - ActionResult.latency_ms: float -> int

M agent/runtime/actions/dispatcher.py
  - started_at=time.time() -> now_iso()
  - finished_at=time.time() -> now_iso()
  - latency_ms=(t1-t0)*1000 -> duration_ms(t1, t0)

M agent/runtime/durable/models.py
  - RuntimeStep.duration_ms: float -> int
  - mark_finished 用 int(round(...))

M agent/runtime/durable/trajectory.py
  - TrajectoryMetrics.duration_ms / TrajectoryRecord.duration_ms: float -> int
  - final_status 算 (ft - st)*1000 -> int(round(...))

M agent/runtime/context_compactor.py
  - CompactionMetric.duration_ms 参数: float -> int
  - 3 处 round(..., 2) -> int(round(...))

M agent/runtime/session_events.py
  - subscribe(timeout: float=25) -> int=25

M agent/modules/remote/core.py
  - DeviceSession.recv(timeout: float=0.5) -> int=0

契约测试 harness/test_v398_type_consistency.py (13 个断言):
- ApprovalRequest.created_at / resolved_at 是 str
- ActionResult.started_at / finished_at 是 str
- ToolDispatcher 路径写 ISO str + int latency_ms
- RuntimeStep / TrajectoryRecord / TrajectoryMetrics duration_ms 是 int
- subscribe / wait / recv 参数 int (用 inspect.signature)
- ToolResult / CompactionMetric duration_ms 是 int
- to_iso/from_iso/duration_ms round-trip 一致
- to_iso 兼容 float 输入 (legacy 调用者)

结果: 8 modified + 2 new = 10 files, +90/-33.
核心契约 505 passed / 0 fail / 4 skipped. 前端 typecheck 0 error.

+++ agent/runtime/session_events.py
@@ -62,7 +62,7 @@ def push_error(session_id: str, error_type: str, message: str):
     push_event(session_id, "error", {"type": error_type, "message": message[:200]})
 
 
-def subscribe(session_id: str, timeout: float = 25) -> Optional[str]:
+def subscribe(session_id: str, timeout: int = 25) -> Optional[str]:
     """Block up to `timeout` seconds for the next SSE-formatted event line.
 
     Returns one SSE frame string, or None if timeout / no session.
